remove multipage activity feature
