+++ packages/doenetml/src/EditorViewer/EditorViewer.tsx
@@ -18,7 +18,7 @@ import { WarningTwoIcon } from "@chakra-ui/icons";
 // @ts-ignore
 import VariantSelect from "./VariantSelect";
 import { CodeMirror } from "@doenet/codemirror";
-import { ActivityViewer } from "../Viewer/ActivityViewer";
+import { PageViewer } from "../Viewer/PageViewer";
 import ErrorWarningPopovers from "./ErrorWarningPopovers";
 import type { WarningDescription, ErrorDescription } from "@doenet/utils";
 import { nanoid } from "nanoid";
@@ -477,7 +477,7 @@ export function EditorViewer({
                     ref={scrollableContainer}
                 >
                     {/* @ts-ignore */}
-                    <ActivityViewer
+                    <PageViewer
                         doenetML={viewerDoenetML}
                         flags={{
                             showCorrectness: true,
@@ -496,15 +496,12 @@ export function EditorViewer({
                         attemptNumber={1}
                         generatedVariantCallback={setVariants}
                         requestedVariantIndex={variants.index}
-                        paginate={paginate}
                         setErrorsAndWarningsCallback={
                             setErrorsAndWarningsCallback
                         }
                         location={location}
                         navigate={navigate}
-                        idsIncludeActivityId={idsIncludeActivityId}
                         linkSettings={linkSettings}
-                        addBottomPadding={addBottomPadding}
                         scrollableContainer={
                             scrollableContainer.current ?? undefined
                         }
