+++ packages/doenetml/src/index.ts
@@ -9,12 +9,6 @@ export {
     // returnAllPossibleVariants,
 } from "@doenet/utils";
 export type { ErrorDescription, WarningDescription } from "@doenet/utils";
-export {
-    calculateOrderAndVariants,
-    determineNumberOfActivityVariants,
-    parseActivityDefinition,
-    returnNumberOfActivityVariantsForCid,
-} from "./utils/activityUtils";
 
 export { CodeMirror } from "@doenet/codemirror";
 export { parseAndCompile } from "@doenet/parser";
